settle down on how to test mobx reactions, started on model test
